swagger and order fix

+++ src/main/kotlin/mimsoft/io/features/cart/CartInfoDto.kt
@@ -6,8 +6,8 @@ data class CartInfoDto(
     val products: List<CartItem> = emptyList(),
     val address : AddressDto? = null,
     val productCount: Int? = null,
-    val productsPrice: Double? = null,
-    val productsDiscount: Double? = null,
-    val totalPrice: Double? = null,
-    val totalDiscount: Double? = null
+    val productsPrice: Long? = null,
+    val productsDiscount: Long? = null,
+    val totalPrice: Long? = null,
+    val totalDiscount: Long? = null
 )
